audio work

+++ src/settings/audio.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Hash)]
 pub struct AudioSettings {
-    pub latency: u16,
+    pub latency: u8,
     pub volume: u8,
 }
